IRIS support for unauthenticated session

+++ scripts/arast.py
@@ -120,7 +120,20 @@ def main():
     oauth_parser.read(oauth_file)
     reauthorize = True
 
+    if "KB_RUNNING_IN_IRIS" in os.environ:
+        if "KB_AUTH_TOKEN" in os.environ and "KB_AUTH_USER_ID" in os.environ and \
+                len(os.environ["KB_AUTH_USER_ID"]) > 0 and \
+                len(os.environ["KB_AUTH_TOKEN"]) > 0 :
+            a_user = os.environ["KB_AUTH_USER_ID"]
+            a_token = os.environ["KB_AUTH_TOKEN"]
+        else:
+            print "Please authenticate with KBase credentials"
+            sys.exit()
+
     if args.command == 'logout' or args.command == 'login':
+        if "KB_RUNNING_IN_IRIS" in os.environ:        
+            print "Please use the IRIS controls to log in/out"
+            sys.exit()
         try:
             os.remove(oauth_file)
         except:
@@ -129,42 +142,36 @@ def main():
             print >> sys.stderr, '[x] Logged out'
             sys.exit()
 
-    if "KB_RUNNING_IN_IRIS" in os.environ:
-        if "KB_AUTH_TOKEN" in os.environ and "KB_AUTH_USER_ID" in os.environ:
-            a_user = os.environ["KB_AUTH_USER_ID"]
-            a_token = os.environ["KB_AUTH_TOKEN"]
+
+    # Check if user file exists
+    if os.path.exists(oauth_file):
+        token_date_str = oauth_parser.get('auth', 'token_date')
+        tdate = datetime.datetime.strptime(token_date_str, '%Y-%m-%d').date()
+        cdate = datetime.date.today()
+        if (cdate - tdate).days > expiration:
+            reauthorize = True
         else:
-            print("Please log in")
+            reauthorize = False
+    if not reauthorize:
+        a_user = oauth_parser.get('auth', 'user')
+        a_token = oauth_parser.get('auth', 'token')
+        # print >> sys.stderr, "Logged in as: {}".format(a_user)
     else:
-        # Check if user file exists
-        if os.path.exists(oauth_file):
-            token_date_str = oauth_parser.get('auth', 'token_date')
-            tdate = datetime.datetime.strptime(token_date_str, '%Y-%m-%d').date()
-            cdate = datetime.date.today()
-            if (cdate - tdate).days > expiration:
-                reauthorize = True
-            else:
-                reauthorize = False
-        if not reauthorize:
-            a_user = oauth_parser.get('auth', 'user')
-            a_token = oauth_parser.get('auth', 'token')
-            # print >> sys.stderr, "Logged in as: {}".format(a_user)
-        else:
-            print("Please authenticate with KBase credentials")
-            a_user = raw_input("KBase Login: ")
-            a_pass = getpass.getpass(prompt="KBase Password: ")
-            globus_map = get_token(a_user, a_pass)
-            a_token = globus_map['access_token']
-            try:
-                os.makedirs(user_dir)
-            except:
-                pass
-            uparse = SafeConfigParser()
-            uparse.add_section('auth')
-            uparse.set('auth', 'user', a_user)
-            uparse.set('auth', 'token', a_token)
-            uparse.set('auth', 'token_date', str(datetime.date.today()))
-            uparse.write(open(oauth_file, 'wb'))
+        print("Please authenticate with KBase credentials")
+        a_user = raw_input("KBase Login: ")
+        a_pass = getpass.getpass(prompt="KBase Password: ")
+        globus_map = get_token(a_user, a_pass)
+        a_token = globus_map['access_token']
+        try:
+            os.makedirs(user_dir)
+        except:
+            pass
+        uparse = SafeConfigParser()
+        uparse.add_section('auth')
+        uparse.set('auth', 'user', a_user)
+        uparse.set('auth', 'token', a_token)
+        uparse.set('auth', 'token_date', str(datetime.date.today()))
+        uparse.write(open(oauth_file, 'wb'))
 
     if args.command == 'login':
         print "Logged in"
